+++ scripts/test/test.mjs
@@ -3,14 +3,14 @@
  * See https://github.com/jsenv/jsenv-core/blob/master/docs/testing/readme.md#jsenv-test-runner
  */
 
-import { executeTestPlan, nodeProcess } from "@jsenv/core"
+import { executeTestPlan, nodeChildProcess } from "@jsenv/core"
 
 await executeTestPlan({
   rootDirectoryUrl: new URL("../../", import.meta.url),
   testPlan: {
     "tests/**/*.test.mjs": {
       node: {
-        runtime: nodeProcess,
+        runtime: nodeChildProcess,
       },
     },
   },